+++ src/app/result/[id]/meal-container.tsx
@@ -1,6 +1,5 @@
 import { MealImage } from '@/src/app/result/[id]/meal-image';
 import { DetailedMealWithTranslations } from '@/src/types/api.types';
-import { MealIngredientContainer } from '@/src/app/result/[id]/meal-ingredient-container';
 import { MealGeneral } from '@/src/app/result/[id]/meal-general';
 import { MealCommentContainer } from '@/src/app/result/[id]/meal-comment-container';
 
@@ -9,13 +8,12 @@ interface MealContainerProps {
 }
 
 export function MealContainer({ complexMealObject }: MealContainerProps) {
-    const { meal, ingredients, recipe } = complexMealObject;
+    const { meal, description, ingredients, recipe } = complexMealObject;
 
     return (
         <>
             <MealImage imgUrl={meal.imgUrl} title={meal.title} provider={meal.provider} />
-            <MealGeneral meal={meal} />
-            <MealIngredientContainer complexMealObject={complexMealObject} />
+            <MealGeneral meal={meal} description={description} ingredients={ingredients} recipe={recipe} />
             <MealCommentContainer />
         </>
     );