+++ src/app/result/[id]/meal-recipe.tsx
@@ -1,32 +1,38 @@
 import styles from '@/styles/app/result/page.module.scss';
-import { DetailedMeal } from '@/src/types/api.types';
+import { DetailedMeal, MealRecipeSection } from '@/src/types/api.types';
 
 interface MealRecipeProps {
     meal: DetailedMeal;
+    recipe?: MealRecipeSection[];
 }
 
-export function MealRecipe({ meal }: MealRecipeProps) {
-    const { recipeSections } = meal;
+export function MealRecipe({ recipe }: MealRecipeProps) {
+    if (!recipe || recipe.length === 0) {
+        return (
+            <div className={styles['instruction-section']}>
+                <h5>Przepis:</h5>
+                <p>Niesety, autor nie dostarczył żadnego przepisu dla tego posiłku.</p>
+            </div>
+        );
+    }
 
     return (
         <div>
-            {recipeSections.length === 0
-                ? 'Unfortunately, the author does not provide any recipe for this meal.'
-                : recipeSections.map(section => {
-                    return (
-                        <div className={styles['instruction-section']}>
-                            <h5>Przepis: {section.name || '-'}</h5>
-                            <ol>
-                                {section.steps.map((step, idx) => {
-                                    // NOTE: Index is okay, because it's a static list
-                                    return (
-                                        <li key={idx}>{step}</li>
-                                    );
-                                })}
-                            </ol>
-                        </div>
-                    );
-                })}
+            {recipe && recipe.map(section => {
+                return (
+                    <div className={styles['instruction-section']}>
+                        <h5>Przepis{section.name ? `na ${section.name}` : ''}:</h5>
+                        <ol>
+                            {section.steps.map((step, idx) => {
+                                // NOTE: Index is okay, because it's a static list
+                                return (
+                                    <li key={idx}>{step}</li>
+                                );
+                            })}
+                        </ol>
+                    </div>
+                );
+            })}
         </div>
     );
 }